+++ decoder_forge/uc_generate_code.py
@@ -1,7 +1,8 @@
+import logging
+import yaml
+
 from decoder_forge.i_printer import IPrinter
 from decoder_forge.i_template_engine import ITemplateEngine
-import logging
-import json
 from decoder_forge.pattern import Pattern
 from decoder_forge.pattern_algorithms import build_decode_tree_by_fixed_bits
 from decoder_forge.pattern_algorithms import flatten_decode_tree
@@ -9,12 +10,12 @@ from decoder_forge.pattern_algorithms import flatten_decode_tree
 logger = logging.getLogger(__name__)
 
 
-def uc_generate_code(printer: IPrinter, tengine: ITemplateEngine, input_json: str):
+def uc_generate_code(printer: IPrinter, tengine: ITemplateEngine, input_yaml: str):
     """
     Generate and output a decoder in an specified language based on patterns provided
-    in a JSON string.
+    in a YAML string.
 
-    This function decodes a JSON input (using a custom object hook) to extract Pattern
+    This function decodes a YAML input (using a custom object hook) to extract Pattern
     objects, builds a hierarchical decode tree, and then flattens this tree for easier
     processing. It uses a template engine (tengine) to generate code based on a context
     that includes the repository mapping, decode tree, and flattened tree. Each line of
@@ -24,15 +25,15 @@ def uc_generate_code(printer: IPrinter, tengine: ITemplateEngine, input_json: st
         printer (IPrinter): An instance of IPrinter responsible for printing output.
         tengine (ITemplateEngine): An instance of ITemplateEngine used for code
            generation.
-        input_json (str): A JSON string representing a list of pattern dictionaries.
+        input_yaml (str): A YAML string representing a list of pattern dictionaries.
 
     Raises:
-        json.JSONDecodeError: If the input JSON is invalid.
+        yaml.YAMLError: If the input YAML is invalid.
         Exception: For any unexpected error during pattern processing or code generation.
 
     Example:
-        >>> json_input = '[{"pattern": "1010", "name": "PatternA"}]'
-        >>> uc_generate_code(printer, tengine, json_input)
+        >>> yaml_input = '[{"pattern": "1010", "name": "PatternA"}]'
+        >>> uc_generate_code(printer, tengine, yaml_input)
     """
 
     def pattern_decoder(dct):
@@ -43,7 +44,16 @@ def uc_generate_code(printer: IPrinter, tengine: ITemplateEngine, input_json: st
         return dct  # fallback to default behavior
 
     logger.info("Call: uc_generate_code")
-    ins = json.loads(input_json, object_hook=pattern_decoder)
+    ins = yaml.load(input_yaml, Loader=yaml.Loader)
+
+    if ins is None:
+        ins = {}
+
+    if "patterns" not in ins:
+        ins["patterns"] = []
+
+    # convert
+    ins = [pattern_decoder(dct) for dct in ins["patterns"]]
 
     # build pattern list and repository
     pats = [i["pattern"] for i in ins]